fix: parameter to forbid industry generation was broken (#37)
The parameter did not have any effect in game. This is now fixed, at
least in my test games it worked as expected.
As a side effect it appears that the game generation may now more often
run into issues of not being able to place all industries at least once,
depending on map size and mountain height.